+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Unreleased
+- Keyboards detect row of rows and unwraps them, allowing for dynamic building of rows
+- Custom inspect for keybord
 
 ## [0.60.0]
 - Clean README.md and separate content in different guides in [./guides](./guides)